Zotero MCP/requirements.txt
mcp[cli]>=1.2.0
pyzotero>=1.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
